Started work on test runner
