+++ lib/index.js
@@ -10,6 +10,7 @@ const options = {
   loglevel,
   packageManager: argv.packageManager,
   exampleDirs: parseExampleDirArgv(argv.exampleDir),
+  test: argv.test,
 };
 
 const command = `node ./index.js --webpack=${argv.webpack} --dependency=${argv.dependency}${argv.packageManager ? ` --package-manager=${argv.packageManager}` : null}`;
