+++ lib/install-webpack-and-dependency/index.js
@@ -1,5 +1,4 @@
-import childProcess from 'child_process';
-import { reduce } from 'lodash';
+import { runCommand } from '../utils';
 import { ROOT_PATH } from '../consts';
 
 /**
@@ -34,53 +33,19 @@ function getCommands(webpackSetup, dependencySetup, packageManager) {
  * @param {string} packageManager - Package manager that should be used (npm or yarn)
  * @returns {Promise} A promise indicating the installation success
  */
-export default function (webpackSetup, dependencySetup, packageManager) {
+export default async function (webpackSetup, dependencySetup, packageManager) {
   const commands = getCommands(webpackSetup, dependencySetup, packageManager);
 
-  return new Promise((resolve, reject) => {
-    // IMPORTANT: ROOT_PATH needs to contain a package.json file
-    // If it doesn't exist, npm will search for the closest package.json and install into this folder
-    childProcess.exec(commands.install, { cwd: ROOT_PATH }, (err, stdout, stderr) => {
-      if (err) {
-        const error = new Error();
-        error.err = ['Error calling install command', err];
-        return reject(error);
-      }
+  // IMPORTANT: ROOT_PATH needs to contain a package.json file
+  // If it doesn't exist, npm will search for the closest package.json and install into this folder
+  const stdout = await runCommand(commands.install, ROOT_PATH);
 
-      if (stderr) {
-        const outputted = stderr.trim().split('\n');
+  if (stdout.indexOf(webpackSetup.toLocalName()) === -1 ||
+    stdout.indexOf(dependencySetup.toLocalName()) === -1) {
+    const error = new Error();
+    error.err = ['Expected versions not in dependency tree', stdout];
+    throw error;
+  }
 
-        // Ignore warnings in stderr (yarn)
-        const errors = reduce(outputted, (validErrors, errLine) => {
-          if (errLine.indexOf('warning') !== 0) {
-            validErrors.push(errLine);
-          }
-          return validErrors;
-        }, []);
-
-        if (errors.length > 0) {
-          const error = new Error();
-          error.err = ['Error output when installing', errors.join('\n')];
-          return reject(error);
-        }
-      }
-
-      if (stdout.indexOf(webpackSetup.toLocalName()) === -1 ||
-          stdout.indexOf(dependencySetup.toLocalName()) === -1) {
-        const error = new Error();
-        error.err = ['Expected versions not in dependency tree', stdout];
-        return reject(error);
-      }
-
-      return childProcess.exec(commands.installDeps, { cwd: dependencySetup.installLocation }, (e) => {
-        if (e) {
-          const error = new Error();
-          error.err = ['Error calling install command for dependency build', e];
-          return reject(error);
-        }
-
-        return resolve();
-      });
-    });
-  });
+  return runCommand(commands.installDeps, dependencySetup.installLocation);
 }
