+++ lib/runner.js
@@ -4,6 +4,7 @@ import getLogger from './logger';
 import generateInstallObjectFor from './generate-install-object';
 import installWebpackAndDependency from './install-webpack-and-dependency';
 import getDependencyExamples from './get-dependency-examples';
+import runDependencyTests from './run-dependency-tests';
 import runDependencyWithWebpack from './run-dependency-with-webpack';
 
 /**
@@ -48,7 +49,8 @@ export default async function (webpackVersion, dependencyVersion, options) {
   const finishGetExamples = new Date().getTime();
   logger.debug(`Finished get examples (${finishGetExamples - startGetExamples}ms)`);
 
-  if (isEmpty(dependencyExamples)) {
+  // Ignore missing examples if we should run tests
+  if (isEmpty(dependencyExamples) && !options.test) {
     const error = new Error();
     error.err = 'Unable to get any dependency examples';
     throw error;
@@ -81,5 +83,11 @@ export default async function (webpackVersion, dependencyVersion, options) {
     throw err;
   }
 
-  return dependencyExamples;
+  if (options.test) {
+    logger.info(`Running ${chalk.bold(webpackSetup)} tests with ${chalk.bold(options.test)}...`);
+    const testResults = await runDependencyTests(webpackSetup, dependencySetup, options);
+    logger.info(testResults);
+  }
+
+  // return dependencyExamples;
 }
